requirements_minimal.txt
tensorflow
numpy   
opencv-python
hydra-core


